Move nmea structures in new files
