+++ src/gpsRtcm.hpp
@@ -29,26 +29,6 @@ struct GpsRtcmData {
 };
 extern GpsRtcmData gpsRtcmData;
 
-
-struct GpsNmeaOutput {
-  bool udpOutput = false;
-  bool btOutput = false;
-  bool serialOutput = false;
-
-  uint lastSent = 0;
-  enum class GpsQuality : uint8_t {
-    none = 0,
-    gps = 1,
-    dgps = 2,
-    floatrtk = 5,
-    fixedrtk = 4,
-  } qpsquality = GpsQuality::none;
-
-  String lastGGA;
-};
-extern GpsNmeaOutput gpsNmeaOutput;
-
-
 void gpsRtcmCreateUdpReceiveHandler();
 void gpsRtcmBtReceiver( void* z );
 void gpsRtcmNtripReceiver( void* z );
